bg-fix(social): fix social authentication
[Finishes #168371600]
