+++ src/components/views/auth/signUp/SignUp.spec.js
@@ -12,12 +12,6 @@ const initialState = {
     loading: false,
     error: {},
   },
-  social: {
-    user: {},
-    isAuthenticated: false,
-    loading: false,
-    error: {},
-  },
 };
 const props = {
   loading: false,
