--- src/redux/actions/types/socialType.js
@@ -1,2 +1,0 @@
-export const SOCIAL_LOGIN_SUCCESS = 'SOCIAL_LOGIN_SUCCESS';
-export const SOCIAL_LOGIN_FAILURE = 'SOCIAL_LOGIN_FAILURE';
