+++ src/redux/reducers/index.js
@@ -2,7 +2,6 @@ import { combineReducers } from 'redux';
 import LandingPageReducer from './landingPageReducer';
 import authReducer from './authReducer';
 import passwordResetReducer from './passwordResetReducer';
-import socialReducer from './socialReducer';
 import singleArticleReducer from './articleReducer';
 
 
@@ -10,6 +9,5 @@ export default combineReducers({
   article: LandingPageReducer,
   auth: authReducer,
   requestReset: passwordResetReducer,
-  social: socialReducer,
   singleArticle: singleArticleReducer,
 });
